requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
pinecone>=6.0.0
langchain==0.3.0
langchain-pinecone==0.2.12
langchain-openai>=0.3.11
langchain-community==0.3.0
sentence-transformers==3.1.1
pypdf==5.1.0
torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu